readme: fix the layout:include example
When I copy-pasted this example as-is into an app to try it out, it didn’t work.
Instead of showing “Message goes here!”, it was showing “My modal content”.
Based on the comments in ticket #5 and my own testing, I believe that this
should have been layout:fragment rather than th:fragment.